# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Genet Open API
  description: |
    Genet OpenAPI for Pod, Deployment, StatefulSet, Service, ConfigMap, and Job operations.
    - All resources use simplified JSON request/response models.
    - Effective namespace is always derived from the API key bound `ownerUser`.
    - Server injects ownership and management labels automatically.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - BearerAuth: []
paths:
  /api/open/pods:
    post:
      summary: Create Pod
      description: |
        Create a Pod using JSON request fields consistent with the web UI "Create Pod" form.
        The effective namespace is derived from API key bound `ownerUser` (user-{identifier}).
      operationId: createPod
      tags: [Pods]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAPIPodCreateRequest"
            example:
              image: "nvidia/cuda:12.0.0-base-ubuntu22.04"
              gpuType: "NVIDIA A100"
              gpuCount: 1
              cpu: "8"
              memory: "32Gi"
              shmSize: "1Gi"
              nodeName: "gpu-node-01"
              gpuDevices: [0]
              name: "train"
              userMounts:
                - hostPath: "/mnt/datasets"
                  mountPath: "/workspace-genet/datasets"
                  readOnly: true
      responses:
        "201":
          description: Pod created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIPodCreateResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Conflict (for example active commit job or name conflict)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List Pods
      description: List Pods for API key bound `ownerUser`.
      operationId: listPods
      tags: [Pods]
      responses:
        "200":
          description: Pod list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIPodList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/open/pods/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Pod ID (same as Pod name)

    get:
      summary: Get Pod
      operationId: getPod
      tags: [Pods]
      responses:
        "200":
          description: Pod details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIPod"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: Update Pod
      description: |
        Update Pod using JSON payload (same fields as create).
        Current behavior is **recreate**: delete old pod, then create a new pod with request fields.
        If `name` is omitted, server tries to infer custom suffix from path `id` and reuse it.
      operationId: updatePod
      tags: [Pods]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAPIPodUpdateRequest"
      responses:
        "201":
          description: Pod recreated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIPodCreateResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflict (for example active commit job)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete Pod
      operationId: deletePod
      tags: [Pods]
      responses:
        "200":
          description: Pod deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/open/statefulsets:
    post:
      summary: Create StatefulSet
      description: |
        Create a StatefulSet using JSON request fields consistent with the web UI multi-pod workload form.
        The effective namespace is derived from API key bound `ownerUser` (user-{identifier}).
      operationId: createStatefulSet
      tags: [StatefulSets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAPIStatefulSetCreateRequest"
            example:
              image: "nvidia/cuda:12.0.0-base-ubuntu22.04"
              gpuType: "NVIDIA A100"
              gpuCount: 1
              replicas: 2
              cpu: "8"
              memory: "32Gi"
              shmSize: "1Gi"
              nodeName: "gpu-node-01"
              name: "train"
      responses:
        "201":
          description: StatefulSet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIStatefulSetCreateResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List StatefulSets
      operationId: listStatefulSets
      tags: [StatefulSets]
      responses:
        "200":
          description: StatefulSet list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIStatefulSetList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/open/statefulsets/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: StatefulSet ID / name

    get:
      summary: Get StatefulSet
      operationId: getStatefulSet
      tags: [StatefulSets]
      responses:
        "200":
          description: StatefulSet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIStatefulSet"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete StatefulSet
      operationId: deleteStatefulSet
      tags: [StatefulSets]
      responses:
        "200":
          description: StatefulSet deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/open/deployments:
    post:
      summary: Create Deployment
      description: |
        Create a Deployment using JSON request fields consistent with the web UI workload form.
        Resource values are per replica; total quota usage equals `replicas * per-replica resources`.
        The effective namespace is derived from API key bound `ownerUser` (user-{identifier}).
      operationId: createDeployment
      tags: [Deployments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAPIDeploymentCreateRequest"
            example:
              image: "nvidia/cuda:12.0.0-base-ubuntu22.04"
              gpuType: "NVIDIA A100"
              gpuCount: 1
              replicas: 2
              cpu: "8"
              memory: "32Gi"
              shmSize: "1Gi"
              nodeName: "gpu-node-01"
              name: "train"
      responses:
        "201":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIDeploymentCreateResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List Deployments
      operationId: listDeployments
      tags: [Deployments]
      responses:
        "200":
          description: Deployment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIDeploymentList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/open/deployments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Deployment ID / name

    get:
      summary: Get Deployment
      operationId: getDeployment
      tags: [Deployments]
      responses:
        "200":
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAPIDeployment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete Deployment
      operationId: deleteDeployment
      tags: [Deployments]
      responses:
        "200":
          description: Deployment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API Key from admin page (preferred) or legacy config.yaml openAPI.apiKeys"
  schemas:
    OpenAPIPodCreateRequest:
      type: object
      description: |
        JSON request for creating a Pod.
        Fields are aligned with the web UI "Create Pod" form.
        All resource values keep Kubernetes quantity format (for example `32Gi`, `500m`).
      properties:
        image:
          type: string
          description: Container image (required). Supports full image reference with optional tag.
          example: "nvidia/cuda:12.0.0-base-ubuntu22.04"
        gpuType:
          type: string
          description: GPU type display name. Optional when `gpuCount` is 0.
          example: "NVIDIA A100"
        gpuCount:
          type: integer
          minimum: 0
          maximum: 8
          description: Requested GPU count. `0` means CPU-only pod.
          example: 1
        cpu:
          type: string
          description: CPU request (UI option value, for example `8`).
          example: "8"
        memory:
          type: string
          description: Memory request in Kubernetes quantity format.
          example: "32Gi"
        shmSize:
          type: string
          description: Shared memory size mounted to `/dev/shm`.
          example: "1Gi"
        nodeName:
          type: string
          description: Optional fixed scheduling node.
          example: "gpu-node-01"
        gpuDevices:
          type: array
          description: |
            Optional explicit GPU device indices on selected node.
            When provided, effective GPU count equals `gpuDevices.length`.
          items:
            type: integer
          example: [0, 1]
        name:
          type: string
          description: |
            Optional custom Pod name suffix (for example `train`, `dev`).
            Server generates full name as `pod-{userIdentifier}-{name|timestamp}`.
          example: "train"
        userMounts:
          type: array
          description: Optional custom hostPath mounts (requires admin enabling user mounts).
          items:
            $ref: "#/components/schemas/OpenAPIUserMount"
      required:
        - image
        - gpuCount
    OpenAPIPodUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/OpenAPIPodCreateRequest"
      description: |
        JSON request for Pod update. Current implementation performs recreate.
        Same fields as create request.
    OpenAPIUserMount:
      type: object
      properties:
        hostPath:
          type: string
          description: Host machine path to mount.
          example: "/mnt/datasets"
        mountPath:
          type: string
          description: Container mount path.
          example: "/workspace-genet/datasets"
        readOnly:
          type: boolean
          description: Whether mount is read-only.
          default: false
      required:
        - hostPath
        - mountPath
    OpenAPIPodCreateResponse:
      type: object
      properties:
        message:
          type: string
          description: Creation result message.
          example: "Pod 创建成功"
        id:
          type: string
          description: Pod ID / name.
          example: "pod-alice-train"
        name:
          type: string
          description: Pod name (same as id).
          example: "pod-alice-train"
    OpenAPIPod:
      type: object
      properties:
        id:
          type: string
          description: Pod ID / name.
        name:
          type: string
          description: Pod name.
        namespace:
          type: string
          description: Namespace of this pod.
        container:
          type: string
          description: Main container name.
        status:
          type: string
          description: Normalized status used by UI.
        phase:
          type: string
          description: Raw Kubernetes Pod phase.
        image:
          type: string
          description: Container image.
        gpuType:
          type: string
          description: GPU type label.
        gpuCount:
          type: integer
          description: Requested GPU count.
        cpu:
          type: string
          description: CPU request.
        memory:
          type: string
          description: Memory request.
        createdAt:
          type: string
          format: date-time
          description: Pod creation timestamp.
        nodeIP:
          type: string
          description: Node IP where pod is scheduled.
        protectedUntil:
          type: string
          format: date-time
          nullable: true
          description: Protection deadline if set.
    OpenAPIPodList:
      type: object
      properties:
        pods:
          type: array
          description: Pod list for current owner user.
          items:
            $ref: "#/components/schemas/OpenAPIPod"
        quota:
          $ref: "#/components/schemas/OpenAPIQuota"
    OpenAPIStatefulSetCreateRequest:
      type: object
      description: |
        JSON request for creating a StatefulSet.
        Resource values are per replica; total quota usage equals `replicas * per-replica resources`.
      properties:
        image:
          type: string
          description: Container image.
          example: "nvidia/cuda:12.0.0-base-ubuntu22.04"
        gpuType:
          type: string
          description: GPU type display name. Optional when `gpuCount` is 0.
          example: "NVIDIA A100"
        gpuCount:
          type: integer
          minimum: 0
          maximum: 8
          description: Requested GPU count per replica.
          example: 1
        replicas:
          type: integer
          minimum: 1
          maximum: 8
          description: StatefulSet replica count.
          example: 2
        cpu:
          type: string
          description: CPU request per replica.
          example: "8"
        memory:
          type: string
          description: Memory request per replica.
          example: "32Gi"
        shmSize:
          type: string
          description: Shared memory size per replica.
          example: "1Gi"
        nodeName:
          type: string
          description: Optional fixed scheduling node for all replicas.
          example: "gpu-node-01"
        name:
          type: string
          description: Optional custom StatefulSet name suffix.
          example: "train"
        userMounts:
          type: array
          description: Optional custom hostPath mounts.
          items:
            $ref: "#/components/schemas/OpenAPIUserMount"
      required:
        - image
        - gpuCount
        - replicas
    OpenAPIStatefulSetCreateResponse:
      type: object
      properties:
        message:
          type: string
          description: Creation result message.
          example: "StatefulSet 创建成功"
        id:
          type: string
          description: StatefulSet ID / name.
          example: "sts-alice-train"
        name:
          type: string
          description: StatefulSet name.
          example: "sts-alice-train"
    OpenAPIStatefulSet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        status:
          type: string
        image:
          type: string
        gpuType:
          type: string
        gpuCount:
          type: integer
        cpu:
          type: string
        memory:
          type: string
        replicas:
          type: integer
        readyReplicas:
          type: integer
        createdAt:
          type: string
          format: date-time
        serviceName:
          type: string
        pods:
          type: array
          items:
            $ref: "#/components/schemas/OpenAPIPod"
    OpenAPIStatefulSetList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OpenAPIStatefulSet"
    OpenAPIDeploymentCreateRequest:
      type: object
      description: |
        JSON request for creating a Deployment.
        Resource values are per replica; total quota usage equals `replicas * per-replica resources`.
      properties:
        image:
          type: string
          description: Container image.
          example: "nvidia/cuda:12.0.0-base-ubuntu22.04"
        gpuType:
          type: string
          description: GPU type display name. Optional when `gpuCount` is 0.
          example: "NVIDIA A100"
        gpuCount:
          type: integer
          minimum: 0
          maximum: 8
          description: Requested GPU count per replica.
          example: 1
        replicas:
          type: integer
          minimum: 1
          maximum: 8
          description: Deployment replica count.
          example: 2
        cpu:
          type: string
          description: CPU request per replica.
          example: "8"
        memory:
          type: string
          description: Memory request per replica.
          example: "32Gi"
        shmSize:
          type: string
          description: Shared memory size per replica.
          example: "1Gi"
        nodeName:
          type: string
          description: Optional fixed scheduling node for all replicas.
          example: "gpu-node-01"
        name:
          type: string
          description: Optional custom Deployment name suffix.
          example: "train"
        userMounts:
          type: array
          description: Optional custom hostPath mounts.
          items:
            $ref: "#/components/schemas/OpenAPIUserMount"
      required:
        - image
        - gpuCount
        - replicas
    OpenAPIDeploymentCreateResponse:
      type: object
      properties:
        message:
          type: string
          description: Creation result message.
          example: "Deployment 创建成功"
        id:
          type: string
          description: Deployment ID / name.
          example: "deploy-alice-train"
        name:
          type: string
          description: Deployment name.
          example: "deploy-alice-train"
    OpenAPIDeployment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        status:
          type: string
        image:
          type: string
        gpuType:
          type: string
        gpuCount:
          type: integer
        cpu:
          type: string
        memory:
          type: string
        replicas:
          type: integer
        readyReplicas:
          type: integer
        createdAt:
          type: string
          format: date-time
        pods:
          type: array
          items:
            $ref: "#/components/schemas/OpenAPIPod"
    OpenAPIDeploymentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OpenAPIDeployment"
    OpenAPIQuota:
      type: object
      properties:
        podUsed:
          type: integer
          description: Used pod count.
        podLimit:
          type: integer
          description: Pod quota limit.
        gpuUsed:
          type: integer
          description: Used GPU count.
        gpuLimit:
          type: integer
          description: GPU quota limit.
...
    Error:
      type: object
      properties:
        error:
          type: string
    Message:
      type: object
      properties:
        message:
          type: string
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Missing or invalid Authorization header"
    Forbidden:
      description: API key lacks required owner binding or scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "forbidden: api key must bind ownerUser"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "resource not found"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"